fix multiclass segmentation

+++ layers_segmentation/utils/data_loading.py
@@ -62,7 +62,7 @@ class BasicDataset(Dataset):
 
         return {
             'image': torch.from_numpy(img).type(torch.FloatTensor),
-            'mask': torch.from_numpy(mask).type(torch.FloatTensor)
+            'mask': torch.from_numpy(mask).type(torch.LongTensor)
         }
